requirements.txt
neat_python==0.92
numpy==1.25.1
pygame==2.5.0
